Fmt and fix feature gating
